chore: release 0.10.0
Bump workspace and Python package versions to 0.10.0, update dependencies including Polars to 0.55.1, and make various fixes and enhancements in the chili ecosystem. Notable changes include adjustments to date arithmetic handling, updates to the `rand` function in `chili-op`, and improvements in Python bindings for date conversion.

+++ crates/chili-py/src/lib.rs
@@ -186,7 +186,7 @@ fn spicy_to_py(py: Python<'_>, obj: SpicyObj) -> PyResult<Py<PyAny>> {
         SpicyObj::String(v) => Ok(v.as_bytes().into_pyobject(py)?.into_any().unbind()),
         SpicyObj::Symbol(v) => Ok(v.into_pyobject(py)?.into_any().unbind()),
         // Date => Python datetime.date
-        SpicyObj::Date(v) => Ok(PyDate::from_timestamp(py, v as i64 * 86400)?
+        SpicyObj::Date(v) => Ok(PyDate::from_timestamp(py, v as f64 * 86400.0)?
             .into_any()
             .unbind()),
         // Time => Python datetime.time
